requirements.txt
torch

transformers
numpy
pandas
pillow
matplotlib

streamlit

python-dotenv
faiss-cpu
